Put EndlessRoadOracle's anonymous-namespace subroutines into an 'internal' namespace (with declarations in an -internal header) to facilitate unit-testing.
